tests/io_utils: move *FaultyDisk function to their only user
Similarly to tests/utils, tests/io_utils is a problematic file as it
lumps together unrelated functions and dumps them in a catch-all package
"tests". This commit moves two of these functions to the only test files
that uses them.
